// src/app/features/user-management/pages/function-permission-group-setting/models/system-user.mode.ts
export interface SystemUser {
  DBSource: string;
  UserId: string;
  UserName: string;
  FuctionGroup: string;
  Bhno: string;
  Role: string;
  Email: string;
  ActivityStartDate: string;
  ActivityEndDate: string;
  Status: string;
  LoginErrors: number;
  ModDate: string;
  ModTime: string;
  ModUser: string;
}